// frontend/src/components/welcomeScreen/WelcomeScreen.module.scss.d.ts
/*
 * {COPYRIGHT-TOP}
 * IBM Confidential
 * (C) Copyright IBM Corp.2024
 * 
 * << 5608-WC0/5608-PC4 >>
 * 
 * All Rights Reserved
 * Licensed Material - Property of IBM
 * The source code for this program is not published or otherwise
 * divested of its trade secrets, irrespective of what has
 * been deposited with the U. S. Copyright Office.
 * 
 * U.S. Government Users Restricted Rights
 * - Use, duplication or disclosure restricted by GSA ADP Schedule Contract with IBM Corp.
 * {COPYRIGHT-END}
 */

declare const styles: {
	modal: string;
	buttonSet: string;
	button: string;
	welcomePopupModal: string;
	modalContentFrame: string;
	chatImage: string;
	customModal: string;
	customModalBackground: string;
	blackModal: string;
	welcomechatHeader: string;
	welcomeMessage: string;
	welcomeModalBotInformation: string;
	buttonSetWelcomeScreen: string;
	logo: string;
	buttonSetText: string;
};
export default styles;